requirements.txt
streamlit==1.39.0
python-docx==1.1.2
pydantic==2.9.2
email-validator==2.2.0
jinja2==3.1.4
requests==2.32.3
